Fix: add data files to MANIFEST.in.
The package_data directive seems to indicate which files
must be installed, not which files are included in sdist.
The thing is that I tested the resulting tarball in the
past and I am pretty sure it included the data files
(otherwise simply installing `base` would not work, e.g.
because it would miss the .sql files).